Create Trie data structure and a search function to find all words in a given list that can be formed from a set of characters
In this code, we implement a Trie data structure to efficiently store a list of words. We also provide a function to find all words that can be formed from a given set of characters. Tries are essential for various string-related operations and are widely used in text processing and search algorithms.